refact: restructure folder component
